# src/app/docs/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Use API to call the music generation service of Suno.ai and easily integrate it into agents like GPTs.",
    "version": "1.1.0"
  },
...
  "paths": {
...
    "/api/get_aligned_lyrics": {
      "get": {
        "summary": "Get lyric alignment.",
        "description": "Get lyric alignment.",
        "tags": ["default"],
        "parameters": [
          {
            "name": "song_id",
            "in": "query",
            "required": true,
            "description": "Song ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/schemas/audio_info"
          }
        }
      }
    },
...
    "/api/upload_audio": {
      "post": {
        "summary": "Upload an audio file to Suno.",
        "description": "Uploads an audio file, completes Suno processing, and returns the uploaded clip data. When processing finishes successfully, the returned `s3_id` is the clip id that can be used in `generate_from_audio`.",
        "tags": ["default"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file to upload."
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Wait until Suno finishes processing the upload.",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "s3_id": {
                      "type": "string",
                      "description": "Generated clip id for the uploaded audio."
                    },
                    "title": {
                      "type": "string"
                    },
                    "image_url": {
                      "type": "string"
                    },
                    "has_vocal": {
                      "type": "boolean"
                    },
                    "display_tags": {
                      "type": "string"
                    },
                    "inferred_description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/move_clip": {
      "post": {
        "summary": "Move clips to a workspace.",
        "description": "Moves one or more clips to a target workspace. The workspace can be identified by id or by exact name.",
        "tags": ["default"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["clip_ids"],
                "properties": {
                  "clip_ids": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      {
                        "type": "string",
                        "description": "Comma-separated clip ids."
                      }
                    ]
                  },
                  "workspace_id": {
                    "type": "string"
                  },
                  "workspace_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspace_id": {
                      "type": "string"
                    },
                    "workspace_name": {
                      "type": "string"
                    },
                    "clip_ids": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/workspace_feed": {
      "get": {
        "summary": "Get the feed for a workspace.",
        "description": "Returns feed/clip information for the requested workspace. The workspace can be identified by id or by exact name.",
        "tags": ["default"],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspace_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspace": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    },
                    "clips": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/audio_info"
                      }
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/persona": {
      "get": {
        "summary": "Get persona information and clips.",
        "description": "Retrieve persona information, including associated clips and pagination data.",
        "tags": ["default"],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Persona ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number (defaults to 1)",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "persona": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Persona ID"
                        },
                        "name": {
                          "type": "string",
                          "description": "Persona name"
                        },
                        "description": {
                          "type": "string",
                          "description": "Persona description"
                        },
                        "image_s3_id": {
                          "type": "string",
                          "description": "Persona image URL"
                        },
                        "root_clip_id": {
                          "type": "string",
                          "description": "Root clip ID"
                        },
                        "clip": {
                          "type": "object",
                          "description": "Root clip information"
                        },
                        "persona_clips": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "clip": {
                                "type": "object",
                                "description": "Clip information"
                              }
                            }
                          }
                        },
                        "is_suno_persona": {
                          "type": "boolean",
                          "description": "Whether this is a Suno official persona"
                        },
                        "is_public": {
                          "type": "boolean",
                          "description": "Whether this persona is public"
                        },
                        "upvote_count": {
                          "type": "integer",
                          "description": "Number of upvotes"
                        },
                        "clip_count": {
                          "type": "integer",
                          "description": "Number of clips"
                        }
                      }
                    },
                    "total_results": {
                      "type": "integer",
                      "description": "Total number of results"
                    },
                    "current_page": {
                      "type": "integer",
                      "description": "Current page number"
                    },
                    "is_following": {
                      "type": "boolean",
                      "description": "Whether the current user is following this persona"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing parameter id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Missing parameter id"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "audio_info": {
        "type": "object",
        "required": [
          "id",
          "title",
          "image_url",
          "lyric",
          "audio_url",
          "video_url",
          "created_at",
          "model_name",
          "status",
          "gpt_description_prompt",
          "prompt",
          "type",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "audio id"
          },
          "title": {
            "type": "string",
            "description": "music title"
          },
          "image_url": {
            "type": "string",
            "description": "music cover image"
          },
          "lyric": {
            "type": "string",
            "description": "music lyric"
          },
          "audio_url": {
            "type": "string",
            "description": "music download url"
          },
          "video_url": {
            "type": "string",
            "description": "Music video download link, can be used to share"
          },
          "created_at": {
            "type": "string",
            "description": "Create time"
          },
          "model_name": {
            "type": "string",
            "description": "suno model name, chirp-v3"
          },
          "status": {
            "type": "string",
            "description": "The generated states include submitted, queue, streaming, complete."
          },
          "gpt_description_prompt": {
            "type": "string",
            "description": "Simple mode on user input prompt, Suno will generate formal prompts, lyrics, etc."
          },
          "prompt": {
            "type": "string",
            "description": "The final prompt for executing the generation task, customized by the user in custom mode, automatically generated by Suno in simple mode."
          },
          "type": {
            "type": "string",
            "description": "Type"
          },
          "tags": {
            "type": "string",
            "description": "Music genre. User-provided in custom mode, automatically generated by Suno in simple mode."
          }
        },
        "title": "audio_info",
        "description": "Audio Info"
      }
    }
  }
}